apps/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
Authlib==1.3.0
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
